% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bicluster_net_methods.R
\name{apply_threshold,bicluster_net-method}
\alias{apply_threshold,bicluster_net-method}
\title{Apply a threshold to a bicluster similarity adjacency matrix.}
\usage{
\S4method{apply_threshold}{bicluster_net}(bic_net)
}
\arguments{
\item{bic_net}{An object of class \code{bicluster_net}.}
}
\value{
An adjacency matrix with the applied threshold.
}
\description{
All values lower than the threshold will be replaced by a 0.
}
\examples{
m <- matrix(rnorm(10000), nrow=100)
bics <- c(run_fabia(m), run_isa(m), run_plaid(m))
bn <- bicluster_network(bics, m)
apply_threshold(bn)

}
