% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{attr_overlap}
\alias{attr_overlap}
\title{Count how often row/column elements occur in biclusters.}
\usage{
attr_overlap(bics, named)
}
\arguments{
\item{bics}{A list of \code{\link{bicluster}} objects.}

\item{named}{Boolean, indicating, if all bicluster objects have names.}
}
\value{
A Data Frame with the counts oof all elements.
}
\description{
Given a list of bicluster objects (\code{\link{bicluster}}), 
the function counts the occurance of all elements in the biclusters.
}
\examples{
m <- matrix(seq(1:16), nrow=4)
# m <- matrix(rnorm(10000), nrow=100)
# bics <- c(run_fabia(m), run_isa(m), run_plaid(m))
# attr_overlap(bics, named=FALSE)

}
