% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bicluster.R
\docType{class}
\name{bicluster-class}
\alias{bicluster-class}
\alias{bicluster}
\title{A S4 class to store biclusters.}
\description{
A S4 class to store biclusters.
}
\section{Slots}{

\describe{
\item{\code{row}}{A vector of row.}

\item{\code{column}}{A vector of columns.}

\item{\code{rowname}}{A vector of names for the rows in \code{row}.}

\item{\code{colname}}{A vector of names for the columns in \code{column}.}

\item{\code{algorithm}}{Algorithm that predicted this bicluster.}
}}

\examples{
bicluster(row=c(1,2), column=c(1,2), 
    rowname=c("a", "b"), colname=c("e", "f"))
    
}
