% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{bicluster_to_matrix}
\alias{bicluster_to_matrix}
\title{Convert a bicluster object to an acutal submatrix of the original matrix.}
\usage{
bicluster_to_matrix(m, bic)
}
\arguments{
\item{m}{Matrix on which the bicluster was computed}

\item{bic}{Bicluster object}
}
\value{
A matrix.
}
\description{
Convert a bicluster object to an acutal submatrix of the original matrix.
}
\examples{
bicluster_to_matrix(matrix(seq(1:16), nrow=4), 
    bicluster(row=c(1,2), column=c(1,2)))
}
