% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{check_names}
\alias{check_names}
\title{Throw an error, if a matrix has not both row- and colnames.}
\usage{
check_names(m)
}
\arguments{
\item{m}{A matrix.}
}
\value{
Throws error, if matrix has no row- and column names.
}
\description{
Throw an error, if a matrix has not both row- and colnames.
}
\examples{
check_names(matrix(c(1,2,3,4), nrow=2))

m <- matrix(c(1,2,3,4), nrow=2)
rownames(m) <- c("r1", "r2")
rownames(m) <- c("c1", "c2")
check_names(m)

}
