% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{clean_bicluster_list}
\alias{clean_bicluster_list}
\title{Clean a list of biclusters, by returning only the valid ones,}
\usage{
clean_bicluster_list(bics)
}
\arguments{
\item{bics}{A list of bicluster objects.}
}
\value{
A lis tof bicluster objects
}
\description{
Clean a list of biclusters, by returning only the valid ones,
}
\examples{
b <- list(bicluster(row=c(1,2,3,4), column=c(1,2,3,4)),
        bicluster(row=c(3,4,5,6), column=c(3,4,5,6)))
clean_bicluster_list(b)

}
