% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bicluster.R
\docType{class}
\name{cooccurrence_net-class}
\alias{cooccurrence_net-class}
\alias{cooccurrence_net}
\title{A S4 class to store co-occurence networks.}
\description{
Object that is returned e.g. be the function \code{\link{feature_network}}.
}
\section{Slots}{

\describe{
\item{\code{adjacency_matrix}}{Adjacency matrix of row- and column-element
co-occurences.}

\item{\code{threshold}}{Estimated threshold for the co-occurence adjacency matrix.
All values lower than that in the matrix should be
discarded. (Note that the indicated threshold is not
applied to the \code{adjacency_matrix})}
}}

\examples{
cooccurrence_net(adjacency_matrix=matrix(seq(1:16), nrow=4),
    threshold=4)
    
}
