% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_matrix_subsetting}
\alias{cpp_matrix_subsetting}
\title{Subsetting of R matrices within c++.}
\usage{
cpp_matrix_subsetting(m, bic)
}
\arguments{
\item{m}{A numeric matrix}

\item{bic}{A bicluster object.}
}
\value{
Matrix subset.
}
\description{
Subsetting of R matrices within c++.
}
\examples{
cpp_matrix_subsetting(matrix(seq(1:16), nrow=4), 
    bicluster(row=c(1,2), column=c(1,2)))
    
}
