% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{detect_elements}
\alias{detect_elements}
\title{Detect the number of elements in a list of biclusters.}
\usage{
detect_elements(bics, MARGIN = "row")
}
\arguments{
\item{bics}{A list of bicluster objects.}

\item{MARGIN}{Choose if the distance is computed over \code{"row"} 
or \code{"column"}.}
}
\value{
Return highest row or column index from a list of biclusters.
}
\description{
Finds the highest element in a list of bicluster objects.
}
\examples{
b <- list(bicluster(row=c(1,2,3,4), column=c(1,2,3,4)),
        bicluster(row=c(3,4,5,6), column=c(3,4,5,6)))
detect_elements(b)

}
