% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{dim,bicluster-method}
\alias{dim,bicluster-method}
\title{Get the dimensions of a bicluster.}
\usage{
\S4method{dim}{bicluster}(x)
}
\arguments{
\item{x}{A bicluster object.}
}
\value{
A numeric vector with the lengths of the rows and
columns of the bicluster.
}
\description{
Get the dimensions of a bicluster.
}
\examples{
dim(bicluster(row=c(1,2), column=c(1,2)))
}
