% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_louvain_overlap.R
\name{feature_louvain_overlap}
\alias{feature_louvain_overlap}
\title{Overlap of features/samples in different louvain communities}
\usage{
feature_louvain_overlap(overlap_tables, mat)
}
\arguments{
\item{overlap_tables}{List of tables as returned by \code{\link{attr_overlap}}.}

\item{mat}{The data matrix used as input for the biclustering algorithms.}
}
\value{
List of tables as returned by \code{\link{attr_overlap}}, 
extended by a column showing how often elements occur across all tables.
}
\description{
The function calculates how often features or samples occur across 
all calculated louvain communities
}
\examples{
# a = data.frame(type=c("row", "row", "row", "column", "column", "column"), 
#   ID=c(1,2,3,1,2,3), Fraction=c(1,1,1,.5, .5, .5))
# b = data.frame(type=c("row", "row", "row", "column", "column", "column"), 
#   ID=c(3,2,4,1,5,3), Fraction=c(1,1,1,.5, .5, .5))
# inl <- list(a, b)
# feature_louvain_overlap(outl, matrix(1:100, nrow=10))

}
