% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{filter_matrix}
\alias{filter_matrix}
\title{Filter a matrix}
\usage{
filter_matrix(mat, threshold = 1)
}
\arguments{
\item{mat}{A Numeric matrix.}

\item{threshold}{All values below will be replaces by 0.}
}
\value{
A filtered numeric matrix.
}
\description{
All values below the threshold will be replaces by 0.
}
\examples{
m <- matrix(rnorm(10000), nrow=100)
filter_matrix(m, threshold=1)

}
