% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bicluster_net_methods.R
\name{get_adjacency}
\alias{get_adjacency}
\title{Get Adjacency matrix}
\usage{
get_adjacency(bic_net)
}
\arguments{
\item{bic_net}{An object of class \code{bicluster_net}.}
}
\value{
Raw unfiltered adjacency matrix.
}
\description{
Return Adjacency matrix from bicluster network
}
\examples{
m <- matrix(rnorm(10000), nrow=100)
bics <- c(run_fabia(m), run_isa(m), run_plaid(m))
bn <- bicluster_network(bics, m)
get_adjacency(bn)

}
