% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get_algorithms}
\alias{get_algorithms}
\title{Get Algorithms}
\usage{
get_algorithms(bics)
}
\arguments{
\item{bics}{a list of \code{\link{bicluster}} objects.}
}
\value{
A character vector with algorithm names
}
\description{
Get a unique vector of algorithms from a
list of \code{\link{bicluster}} objects.
}
\examples{
b <- list(bicluster(row=c(1,2,3,4), column=c(1,2,3,4), algorithm="isa"),
    bicluster(row=c(3,4,5,6), column=c(3,4,5,6), algorithm="QUBIC"))
}
