% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bicluster_net_methods.R
\name{get_bic_net_algorithms,bicluster_net-method}
\alias{get_bic_net_algorithms,bicluster_net-method}
\title{Get bicluster network algorithms}
\usage{
\S4method{get_bic_net_algorithms}{bicluster_net}(bic_net)
}
\arguments{
\item{bic_net}{An object of class \code{bicluster_net}.}
}
\value{
Algorithm names as characters.
}
\description{
Return algorithms from bicluster network
}
\examples{
m <- matrix(rnorm(10000), nrow=100)
bics <- c(run_fabia(m), run_isa(m), run_plaid(m))
bn <- bicluster_network(bics, m)
get_bic_net_algorithms(bn)

}
