% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_algorithms.R
\name{get_runibic_function}
\alias{get_runibic_function}
\title{Run the UniBic biclustering algorithm}
\usage{
get_runibic_function()
}
\description{
The function executes the \link[runibic]{BCUnibic} biclustering algorithm,
returning a list of biclusters converted into bicluster objects
compatible with this package.
If the algorithm fails to run, an empty list is returned.
}
\details{
Due to inavailability of runibic on windows, 
execute the result of the function in R to run the UniBic in R.


# @param data_matrix A numeric matrix.

# @param minRow Same parameters as in \link{filter_bicluster_size}.

# @param minCol Same parameters as in \link{filter_bicluster_size}.

# @param ... Other parameters forwarded to the \link[runibic]{BCUnibic}
function.

# @return a list of \link{bicluster} objects.
}
\examples{
m <- matrix(seq(1:16), nrow=4)
# m <- matrix(rnorm(10000), nrow=100)
# run_unibic(m, nbic=10)

}
