% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getallBFClusters}
\alias{getallBFClusters}
\title{Get all biclusters from a Bi-Force output file.}
\usage{
getallBFClusters(filename)
}
\arguments{
\item{filename}{Name of the Bi-Force output file.}
}
\value{
List of biclusters in the form of \code{\link{getBFCluster}}
}
\description{
Get all biclusters from a Bi-Force output file.
}
\examples{
getallBFClusters("PathToBiForceOutput.txt")


}
