% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getBFCluster}
\alias{getBFCluster}
\title{Get a bicluster a Bi-Force output file}
\usage{
getBFCluster(filename, cluster)
}
\arguments{
\item{filename}{Name of the Bi-Force output file.}

\item{cluster}{Number of the bicluster that should be extracted.}
}
\value{
Bicluster as list with rownames in attribute "row" 
and colnames in attribute "column".
}
\description{
Get a bicluster a Bi-Force output file
}
\examples{
a <- "PathToBiForceOutput.txt"
# getBFCluster(a, cluster=1)

}
