% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{has_names}
\alias{has_names}
\title{Check, whether a matrix has row- and colnames.}
\usage{
has_names(m)
}
\arguments{
\item{m}{A matrix}
}
\value{
Logical indicating existence of row- and colnames.
}
\description{
Check, whether a matrix has row- and colnames.
}
\examples{
has_names(matrix(c(1,2,3,4), nrow=2))

m <- matrix(c(1,2,3,4), nrow=2)
rownames(m) <- c("r1", "r2")
rownames(m) <- c("c1", "c2")
has_names(m)

}
