% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mouse_data.R
\docType{data}
\name{mouse_data}
\alias{mouse_data}
\title{Mouse brain lipidomics data}
\format{
A data frame with 245 rows and 61 columns
}
\source{
\url{https://www.ebi.ac.uk/metabolights/MTBLS562}
}
\usage{
mouse_data
}
\description{
Development of the mice brain lipidome over several weeks.
}
\examples{
# data(mouse_data)
}
\keyword{datasets}
