% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{NoBFBiclusters}
\alias{NoBFBiclusters}
\title{Get the number of biclusters, generated by the Bi-Force algorithm.}
\usage{
NoBFBiclusters(filename)
}
\arguments{
\item{filename}{Name of the Bi-Force output file.}
}
\value{
Number of biclusters.
}
\description{
Get the number of biclusters, generated by the Bi-Force algorithm.
}
\examples{
# NoBFBiclusters("PathToBiForceOutput.txt")

}
