% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{occurance_matrix}
\alias{occurance_matrix}
\title{Occurance matrix of data points in a list of biclusters}
\usage{
occurance_matrix(bics, mat)
}
\arguments{
\item{bics}{A list of \code{\link{bicluster}} objects.}

\item{mat}{The data matrix used for biclustering.}
}
\value{
A numeric matrix with the dimensions of the input matrix. 
The values represent the frequency of occurance of this point in 
the list of biclusters.
}
\description{
The function computes a matrix with the same dimensions as the input 
matrix and fills the matrix elements with the frequence of occurance 
of the data points in the input list of biclusters.
}
\examples{
m <- matrix(rnorm(10000), nrow=100)
bics <- c(run_fabia(m), run_isa(m), run_plaid(m))
occurance_matrix(bics, m)

}
