% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{p_overlap}
\alias{p_overlap}
\title{Probability for an overlap of two samples.}
\usage{
p_overlap(x, y, k, n)
}
\arguments{
\item{x}{Overlap.}

\item{y}{Size of sample 1.}

\item{k}{Size of Sample 2.}

\item{n}{Number of all elements sampled from.}
}
\value{
Overlap probability.
}
\description{
The probability is computed using the 
forumla \eqn{\frac{{y \choose x}\times {n-y \choose k-x}}{{n \choose k}}}.
}
\examples{
p_overlap(10, 20, 30, 100)

}
