% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{p_overlap_2d}
\alias{p_overlap_2d}
\title{Probability for an overlap of two dimensional samples}
\usage{
p_overlap_2d(ov_x, ov_y, s1x, s1y, s2x, s2y, mat_x, mat_y)
}
\arguments{
\item{ov_x}{Overlap in the first dimension.}

\item{ov_y}{Overlap in the second dimension.}

\item{s1x}{First sample of the first dimension.}

\item{s1y}{First sample of the second dimension.}

\item{s2x}{Second sample of first dimension.}

\item{s2y}{Second sample of the second dimension.}

\item{mat_x}{Number of all elements from the first 
dimension sampled from.}

\item{mat_y}{Number of all elements from the second 
dimension sampled from.}
}
\value{
Overlap probability.
}
\description{
Is computed by calculating the overlap probability for each 
dimension independently and multiplying them using the 
function \code{\link{p_overlap}}.
}
\examples{
p_overlap_2d(10, 10, 20, 20, 30, 30, 100, 100)

}
