% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{p_overlap_higher}
\alias{p_overlap_higher}
\title{Probability for an overlap higher or equal to the observed 
one of two samples}
\usage{
p_overlap_higher(x, y, k, n)
}
\arguments{
\item{x}{Overlap.}

\item{y}{Size of sample 1.}

\item{k}{Size of Sample 2.}

\item{n}{Number of all elements sampled from.}
}
\value{
Overlap probability.
}
\description{
Is computed by adding up probabilities for all possible 
overlaps equal or higher to the observed one using the 
function \code{\link{p_overlap}}.
}
\examples{
p_overlap_higher(10, 20, 30, 100)

}
