% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bicluster_net_methods.R
\name{plot,bicluster_net,missing-method}
\alias{plot,bicluster_net,missing-method}
\title{Plot a bicluster network}
\usage{
\S4method{plot}{bicluster_net,missing}(x, y, ...)
}
\arguments{
\item{x}{An object of class \code{\link{bicluster_net}}.}

\item{y}{Not used.}

\item{...}{Plot parameters forwarded
to \code{igraph::\link[igraph]{plot.igraph}}}
}
\value{
An \link[igraph]{graph} plot.
}
\description{
Converts the object into a \link[igraph]{graph} and uses its plot function.
}
\examples{
# m <- matrix(rnorm(10000), nrow=100)
# bics <- c(run_fabia(m), run_isa(m), run_plaid(m))
# bn <- bicluster_network(bics, m)
# plot(bn)

}
