% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bicluster_net_methods.R
\name{plot,cooccurrence_net,missing-method}
\alias{plot,cooccurrence_net,missing-method}
\title{Plot a co-occurrence network}
\usage{
\S4method{plot}{cooccurrence_net,missing}(x, y, ...)
}
\arguments{
\item{x}{An object of class \code{\link{cooccurrence_net}}.}

\item{y}{Not used.}

\item{...}{Plot parameters forwarded
to \code{igraph::\link[igraph]{plot.igraph}}}
}
\value{
An \link[igraph]{graph} plot.
}
\description{
Converts the object into a \link[igraph]{graph} and uses its plot function.
}
\examples{
m <- matrix(rnorm(10000), nrow=100)
# bics <- c(run_fabia(m), run_isa(m), run_plaid(m))
# fn <- feature_network(bics, m)
# plot(fn)

}
