% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{randomize_matrix}
\alias{randomize_matrix}
\title{Randomize a matrix}
\usage{
randomize_matrix(m)
}
\arguments{
\item{m}{A matrix.}
}
\value{
A randomized version of the input matrix.
}
\description{
Randomize a matrix bu shuffling all rows and columns.
}
\examples{
m <- matrix(c(1,2,3,4), nrow=2)
randomize_matrix(m)
}
