% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{replace_threshold}
\alias{replace_threshold}
\title{Replace elements of an integer matrix.}
\usage{
replace_threshold(m, threshold)
}
\arguments{
\item{m}{A numeric matrix.}

\item{threshold}{A numeric threshold under which all elements in the 
matrix are replaced by zero.}
}
\value{
An integer matrix.
}
\description{
This function replaces all elements of an integer matrix, which are under a 
certain threshold (<) with zero.
}
\examples{
# m <- matrix(rnorm(10000), nrow=100)
# replace_threshold(m, 1)

}
