% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{replace_values}
\alias{replace_values}
\title{Replace values in an integer adjacency matrix.}
\usage{
replace_values(mat, threshold, replace_higher = TRUE)
}
\arguments{
\item{mat}{An integer matrix}

\item{threshold}{All values in the matrix lower than this values are 
replaced by 0.}

\item{replace_higher}{If set to true, all values >= \code{threshold} 
are replaced by 1.}
}
\value{
An integer matrix with (partially) replaced values.
}
\description{
Replace values in an integer matrix, that are lower than a certain 
threshold.
}
\examples{
replace_values(matrix(seq(1, 16), nrow=4), threshold=4)

}
