% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rowhistogram}
\alias{rowhistogram}
\title{Get the rowlengths for a list of bicluster objects.}
\usage{
rowhistogram(bic)
}
\arguments{
\item{bic}{A list  of bicluster objects.}
}
\value{
A vector with the lenghts of the rows in every bicluster object.
}
\description{
Can be used for e.g. histograms.
}
\examples{
m <- matrix(rnorm(10000), nrow=100)
bics <- c(run_fabia(m), run_isa(m), run_plaid(m))
rowhistogram(bics)

}
