% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_algorithms.R
\name{run_akmbiclust}
\alias{run_akmbiclust}
\title{Run the akmbiclust biclustering algorithm}
\usage{
run_akmbiclust(data_matrix, minRow = 2, minCol = 2, ...)
}
\arguments{
\item{data_matrix}{A numeric matrix.}

\item{minRow}{Same parameters as in \link{filter_bicluster_size}.}

\item{minCol}{Same parameters as in \link{filter_bicluster_size}.}

\item{...}{Other parameters forwarded
to the \link[akmbiclust]{akmbiclust} function.}
}
\value{
a list of \link{bicluster} objects.
}
\description{
The function executes the \link[akmbiclust]{akmbiclust} biclustering
algorithm, returning a list of biclusters converted into bicluster
objects compatible with this package.
If the algorithm fails to run, an empty list is returned.
}
\examples{
set.seed(10)
m <- matrix(rnorm(10000), nrow=100)
# Not run: run_akmbiclust(m, k=10)

}
