% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_algorithms.R
\name{run_unibic}
\alias{run_unibic}
\title{Run the UniBic biclustering algorithm}
\usage{
run_unibic(data_matrix, minRow = 2, minCol = 2, ...)
}
\value{
Function as a string, which can be executed.
}
\description{
The function executes the runibic::BCUnibic biclustering algorithm,
returning a list of biclusters converted into bicluster objects
compatible with this package.
If the algorithm fails to run, an empty list is returned.
}
\details{
# @param data_matrix A numeric matrix.

# @param minRow Same parameters as in \link{filter_bicluster_size}.

# @param minCol Same parameters as in \link{filter_bicluster_size}.

# @param ... Other parameters forwarded to the runibic::BCUnibic function.
function.

# @return a list of \link{bicluster} objects.
}
\examples{
m <- matrix(seq(1:16), nrow=4)
# m <- matrix(rnorm(10000), nrow=100)
# run_unibic(m, nbic=10)

}
