% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bicluster_net_methods.R
\name{select_biclusters_from_bicluster_network}
\alias{select_biclusters_from_bicluster_network}
\title{Create a subset of biclusters based on a bicluster network}
\usage{
select_biclusters_from_bicluster_network(bic_net, bics)
}
\arguments{
\item{bic_net}{A \code{\link{bicluster_net}}.}

\item{bics}{A list of \code{\link{bicluster}} objects as returned
by \code{\link{get_biclusters}}.}
}
\value{
A subsetted list of \code{\link{bicluster}} objects
}
\description{
The function returns an adapted bicluster list based on
a \code{\link{bicluster_net}} object.
This might be necessary e.g. after \code{\link{get_louvain_communities}}
was used a community consists only of a subset of the biclusters.
}
\examples{
m <- matrix(rnorm(10000), nrow=100)
bics <- c(run_fabia(m), run_isa(m), run_plaid(m))
bn <- bicluster_network(bics, m)
lc <- get_louvain_communities(bn)
select_biclusters_from_bicluster_network(lc[[1]], bics)

}
