% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{transpose_bicluster}
\alias{transpose_bicluster}
\title{Transpose a bicluster.
Row and column slots will be changed.}
\usage{
transpose_bicluster(bic)
}
\arguments{
\item{bic}{A bicluster object.}
}
\value{
A transposed bicluster object,
}
\description{
Transpose a bicluster.
Row and column slots will be changed.
}
\examples{
transpose_bicluster(bicluster(row=c(3,4,5,6), column=c(3,4,5,6)))

}
