% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{validate_bicluster}
\alias{validate_bicluster}
\title{Indicates, whether a bicluster is valid.
That means it needs at least one row and one column.}
\usage{
validate_bicluster(bic, minRow = 1L, minCol = 1L)
}
\arguments{
\item{bic}{A bicluster object}

\item{minRow}{Minimum number of required rows (Min=1).}

\item{minCol}{Minimum number of required columns (Min=1).}
}
\value{
Logical indicating a valid bicluster object.
}
\description{
Indicates, whether a bicluster is valid.
That means it needs at least one row and one column.
}
\examples{
validate_bicluster(bicluster(row=c(3,4,5,6), column=c(3,4,5,6)))

}
