% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{write_matrix}
\alias{write_matrix}
\title{Write an R matrix to a file (In a Bi-Force or QUBIC2 readable format).}
\usage{
write_matrix(m, filename, qubic2_format = FALSE)
}
\arguments{
\item{m}{A Numeric matrix.}

\item{filename}{Name of the output file.}

\item{qubic2_format}{Write the matrix in a format QUBIC2 is able to read. 
This means adding a row- and column names to the file.}
}
\value{
0 if file was written successfully.
}
\description{
Write an R matrix to a file (In a Bi-Force or QUBIC2 readable format).
}
\examples{
write_matrix(matrix(c(1,2,3,4), nrow=2), "testfile.txt")

}
