% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{zero_subsetting}
\alias{zero_subsetting}
\title{Make a vector of R indices compatible with c++ by 
substracting every element by one.}
\usage{
zero_subsetting(v)
}
\arguments{
\item{v}{A numeric vector.}
}
\value{
A numeric vector with every element decremented by one.
}
\description{
Make a vector of R indices compatible with c++ by 
substracting every element by one.
}
\examples{
zero_subsetting(c(1,2,3,4,5))

}
