% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_algorithms.R
\name{run_unibic}
\alias{run_unibic}
\title{Run the UniBic biclustering algorithm}
\usage{
run_unibic(data_matrix, minRow = 2, minCol = 2, ...)
}
\arguments{
\item{data_matrix}{A numeric matrix.}

\item{minRow}{Same parameters as in \link{filter_bicluster_size}.}

\item{minCol}{Same parameters as in \link{filter_bicluster_size}.}

\item{...}{Other parameters forwarded to the \link[runibic]{BCUnibic}
function.}
}
\value{
a list of \link{bicluster} objects.
}
\description{
The function executes the \link[runibic]{BCUnibic} biclustering algorithm,
returning a list of biclusters converted into bicluster objects
compatible with this package.
If the algorithm fails to run, an empty list is returned.
}
\examples{
m <- matrix(rnorm(10000), nrow=100)
# run_unibic(m, nbic=10)

}
